requirements.txt
PySide6==6.3.1
PySide6-Addons==6.3.1
PySide6-Essentials==6.3.1
python-rtmidi==1.4.9
shiboken6==6.3.1
